Adds modules for downloading files and searching strings
Enhances the UI with visual feedback and improves the user experience by allowing file downloads and string searches within the application.

Also includes a script to create symbol debug files for openMSX,
adjusts the Makefile to incorporate the new modules and support debugging,
and fixes some const definitions in the hget library.