Adds handler-cancellable abort functionality
Introduces a mechanism to allow the user's handler
functions to cancel the hget operation.

This change introduces a new global variable
`cancelled_by_handler` that is set to true when the
`hgetcancel` function is called. This variable is then
checked in the `AbortIfEscIsPressed` macro, which now
also aborts if `cancelled_by_handler` is true. This
allows the user's handler function to cancel the hget
operation by calling `hgetcancel`.