Improves menu navigation and adds VGM disable option
Adds support for shift key to jump to the beginning/end of the file list.

Adds a compilation option to disable VGM support to reduce code size and memory footprint on systems where VGM playback is not required.
This is controlled by the `DISABLE_VGM` preprocessor definition.

Adjusts the help screen to reflect the new navigation feature and possible VGM disable option.